Moved util functions to their own file and exported them in index
